% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{predict}
\alias{predict}
\title{Predict matchings on test data.}
\usage{
predict(
  test_files,
  trained_params,
  target_ess = 100,
  seed = 123,
  use_spf = TRUE,
  parallelize = TRUE
)
}
\arguments{
\item{test_files}{A string vector. List of CSV filenames containing the test data.}

\item{trained_params}{A dataframe. Output of predict().}

\item{target_ess}{An integer scalar. The minimum number of particles per test sample. Default 100.}

\item{seed}{An integer scalar. The seed to use for the random number generator. Default 123.}

\item{use_spf}{A logical scalar. Should we use a streaming particle filter process? Default TRUE.}

\item{parallelize}{A logical scalar. Should the experiment be run with parallelism? Default TRUE.}
}
\value{
A dataframe containing the predicted matchings for each test dataset.
}
\description{
Predict matchings on test data.
}
