\name{sgmwsmcpp-package}
\alias{sgmwsmcpp-package}
\alias{sgmwsmcpp}
\docType{package}
\title{
\packageTitle{sgmwsmcpp}
}
\description{
\packageDescription{sgmwsmcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sgmwsmcpp}
\packageIndices{sgmwsmcpp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{sgmwsmcpp}

Maintainer: \packageMaintainer{sgmwsmcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
